# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WHMS Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "schemas": {
      "UserBase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserBase"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "user"
        ]
      },
      "RateLimitResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              429
            ]
          },
          "message": {
            "type": "string"
          },
          "retryAfter": {
            "type": "number"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "RegisterInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "isImpersonation": {
            "type": "boolean"
          },
          "impersonatorId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "token",
          "expiresAt"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserBase"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshInput": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "GenericSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
      "LogoutInput": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          }
        }
      },
      "EmailSendInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "EmailVerifyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
      "ImpersonationStartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "accessToken": {
            "type": "string",
            "example": "access_token_here"
          },
          "refreshToken": {
            "type": "string",
            "example": "refresh_token_here"
          },
          "sessionId": {
            "type": "string",
            "example": "sess_123"
          },
          "targetUser": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "user_567"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "client@example.com"
              }
            },
            "required": [
              "id",
              "email"
            ]
          }
        },
        "required": [
          "success",
          "accessToken",
          "refreshToken",
          "sessionId",
          "targetUser"
        ]
      },
      "ImpersonationStartInput": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "example": "user_567"
          },
          "reason": {
            "type": "string",
            "example": "Customer support debugging"
          }
        },
        "required": [
          "targetUserId"
        ]
      },
      "ImpersonationStopResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "ImpersonationStopInput": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "example": "sess_123"
          }
        },
        "required": [
          "sessionId"
        ]
      },
...
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user (returns tokens)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh & rotate token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout and terminate session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/send": {
      "post": {
        "summary": "Send verification email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/verify": {
      "get": {
        "summary": "Verify email using token",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "token",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/api/auth/impersonate/start": {
      "post": {
        "summary": "Start impersonation (admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonationStartInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Impersonation started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpersonationStartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/impersonate/stop": {
      "post": {
        "summary": "Stop impersonation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonationStopInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stopped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpersonationStopResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}